move ShotItem to types
